Added test runner for Arduino
The Test.h file is generated on Mac. The .ino file includes Test.h and uploads it to EEPROM and then runs it. To avoid unnecessary EEPROM writes, first check to see if the correct code is already there. Now all the tests run on both Mac and Arduino. But many more tests to run.