Fixed how pointers work
Now when you have a pointer and you assign a pointer to it, you get simple assign and the var has a pointer in it. If you assign a non-pointer to that var, it loads the pointer address and stores the value there. This is implicit dereferencing, which seems to work fine for now. I could later add '*' to do explicit dereferencing if it seems to need to be more clear. But I'd rather not do it. The syntax looks messy. Also, if you pass a value that is a pointer it will deref. This will probably break passing a pointer to a function like Animate.